+++ lib/src/orchestrator/states/preparing_camera_state.dart
@@ -198,6 +198,7 @@ class PreparingCameraState extends CameraState {
       captureMode: nextCaptureMode,
       exifPreferences: cameraContext.exifPreferences,
       videoOptions: saveConfig?.videoOptions,
+      mirrorFrontCamera: saveConfig?.mirrorFrontCamera ?? false,
     );
     _isReady = true;
     return _isReady;