+++ lib/src/widgets/camera_awesome_builder.dart
@@ -220,7 +220,6 @@ class CameraAwesomeBuilder extends StatefulWidget {
           sensorConfig: sensorConfig ??
               SensorConfig.single(
                 sensor: Sensor.position(SensorPosition.back),
-                mirrorFrontCamera: mirrorFrontCamera,
               ),
           enablePhysicalButton: enablePhysicalButton,
           progressIndicator: progressIndicator,